# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "eBay Sales Tool API",
    "version": "1.0.0",
    "description": "API documentation for the eBay Sales Tool"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/listings": {
      "get": {
        "summary": "Retrieve all listings",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by listing status"
          },
          {
            "in": "query",
            "name": "itemId",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by item ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Page size for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of all listings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "properties": {
                    "listings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Listing"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "The listing ID"
                      },
                      "title": {
                        "type": "string",
                        "description": "The title of the listing"
                      },
                      "price": {
                        "type": "number",
                        "description": "The price of the listing"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving listings"
          }
        }
      },
      "post": {
        "summary": "Create a new listing",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Listing"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Listing created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Error creating listing"
          }
        }
      }
    },
    "/api/listings/search": {
      "get": {
        "summary": "Search/filter listings",
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "Search by title"
          },
          {
            "in": "query",
            "name": "minPrice",
            "schema": {
              "type": "number"
            },
            "description": "Minimum price"
          },
          {
            "in": "query",
            "name": "maxPrice",
            "schema": {
              "type": "number"
            },
            "description": "Maximum price"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of listings matching search",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Listing"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error searching listings"
          }
        }
      }
    },
    "/api/listings/{id}": {
      "get": {
        "summary": "Get listing by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Listing found"
          },
          "404": {
            "description": "Listing not found"
          },
          "500": {
            "description": "Error fetching listing"
          }
        }
      },
      "put": {
        "summary": "Update listing by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Listing ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Listing"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Listing updated"
          },
          "404": {
            "description": "Listing not found"
          },
          "500": {
            "description": "Error updating listing"
          }
        }
      },
      "delete": {
        "summary": "Delete listing by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Listing deleted successfully"
          },
          "404": {
            "description": "Listing not found"
          },
          "500": {
            "description": "Error deleting listing"
          }
        }
      }
    },
    "/api/listings/{id}/details": {
      "get": {
        "summary": "Get comprehensive listing details including change history",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" }, "description": "Listing ID" },
          { "in": "query", "name": "history_limit", "schema": { "type": "integer" }, "description": "Optional requested number of history rows (capped by config history_display_limit)" },
          { "in": "query", "name": "history_offset", "schema": { "type": "integer" }, "description": "Optional offset for history pagination (default 0)" }
        ],
        "responses": {
          "200": {
            "description": "Listing details payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "listing": { "$ref": "#/components/schemas/Listing" },
                    "catalog": { "$ref": "#/components/schemas/Catalog" },
                    "sales": { "type": "array", "items": { "$ref": "#/components/schemas/Sale" } },
                    "ownerships": { "type": "array", "items": { "$ref": "#/components/schemas/Ownership" } },
                    "change_history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "action": { "type": "string", "enum": ["create","update","delete"] },
                          "changed_fields": { "type": "array", "items": { "type": "string" } },
                          "created_at": { "type": "string", "format": "date-time" },
                          "before_data": { "type": "object" },
                          "after_data": { "type": "object" }
                        }
                      }
                    },
                    "change_history_total": { "type": "integer" },
                    "change_history_limit": { "type": "integer", "description": "Configured max history_display_limit" },
                    "change_history_requested_limit": { "type": "integer", "nullable": true },
                    "change_history_effective_limit": { "type": "integer" },
                    "change_history_offset": { "type": "integer" }
                  }
                }
              }
            }
          },
          "404": { "description": "Listing not found" },
          "500": { "description": "Error fetching listing details" }
        }
      }
    },
...
  },
...
}